backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
ultralytics
opencv-python-headless
albumentations
numpy
Pillow
